+++ lib/apns.dart
@@ -1,8 +1,10 @@
 import 'dart:io';
 
-import 'package:flutter_apns/apns_connector.dart';
-import 'package:flutter_apns/connector.dart';
-import 'package:flutter_apns/firebase_connector.dart';
+import 'package:flutter_apns/src/apns_connector.dart';
+import 'package:flutter_apns/src/connector.dart';
+import 'package:flutter_apns/src/firebase_connector.dart';
+
+export 'package:flutter_apns/src/connector.dart';
 
 /// Creates either APNS or Firebase connector to manage the push notification registration.
 PushConnector createPushConnector() {
