+++ lib/flutter_apns.dart
@@ -1,0 +1,2 @@
+export 'package:flutter_apns/src/connector.dart';
+export 'package:flutter_apns/apns.dart';
