+++ lib/src/apns_connector.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:flutter_apns/connector.dart';
+import 'package:flutter_apns/src/connector.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart' hide MessageHandler;
-export 'package:flutter_apns/connector.dart';
+export 'package:flutter_apns/src/connector.dart';
 
 class ApnsPushConnector extends PushConnector {
   final MethodChannel _channel = const MethodChannel('flutter_apns');
